medsiglip_inference_endpoint/requirements.txt
torch
transformers
pillow
sentencepiece
torchvision